requirements.txt
azure.cognitiveservices.speech
audioplayer
requests
soundfile
mutagen
pytchat
flask
flask_socketio
pytz